+++ src/control_plane/shared_infrastructure/postgres/mod.rs
@@ -1,0 +1,5 @@
+mod postgres_logical_resource_plan;
+mod postgres_plan_error;
+
+pub(crate) use postgres_logical_resource_plan::PostgresLogicalResourcePlan;
+pub(crate) use postgres_plan_error::PostgresPlanError;